Added roles for posts

+++ admin_panel_vlog/app/Http/Controllers/PostController.php
@@ -4,10 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Post;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+//        $this->authorizeResource(Post::class, 'post');
+    }
+
     public function index()
     {
         $posts = Post::all();
@@ -16,6 +23,10 @@ class PostController extends Controller
 
     public function create()
     {
+        if (Gate::denies('create-post')) {
+            abort(403);
+        }
+
         $categories = Category::all();
         $visibilityOptions = ['PUBLIC' => 'Public', 'PRIVATE' => 'Private'];
         return view('posts.create', compact('categories', 'visibilityOptions'));
@@ -33,7 +44,7 @@ class PostController extends Controller
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->visibility = $request->input('visibility');
-        $post->user_id = 1;
+        $post->user_id = 2;
 
         if ($request->hasFile('image')) {
             $imagePath = $request->file('image')->store('uploads/posts', 'public');
@@ -55,7 +66,15 @@ class PostController extends Controller
 
     public function edit($id)
     {
+        $user = User::findOrFail(3);
+
+//        dd($user->isModerator(), $user->id);
         $post = Post::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403);
+        }
+
         $categories = Category::all();
         $visibilityOptions = ['PUBLIC' => 'Public', 'PRIVATE' => 'Private'];
         return view('posts.edit', compact('post', 'categories', 'visibilityOptions'));
@@ -64,6 +83,10 @@ class PostController extends Controller
     public function update(Request $request, $id)
     {
         $post = Post::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403);
+        }
 
         $validatedData = $request->validate([
             'title' => 'required|max:255',
@@ -74,7 +97,7 @@ class PostController extends Controller
 
         $post->title = $validatedData['title'];
         $post->content = $validatedData['content'];
-        $post->visibility = $validatedData['visibility'];
+        $post->visibility = $request->input('visibility');
 
         if ($request->hasFile('image')) {
             $imagePath = $request->file('image')->store('uploads/posts', 'public');
@@ -91,6 +114,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
+
+        if (Gate::denies('delete-post')) {
+            abort(403);
+        }
+
         $post->delete();
         return redirect()->route('posts.index');
     }
